5.4 bootup fix: Variable declaration types

+++ kernel/sched/core.c
@@ -2775,8 +2775,8 @@ int wake_up_state(struct task_struct *p, unsigned int state)
 }
 
 #ifdef CONFIG_SCHED_BORE
-extern u8   sched_burst_fork_atavistic;
-extern uint sched_burst_cache_lifetime;
+extern int sched_burst_fork_atavistic;
+extern int sched_burst_cache_lifetime;
 
 static void __init sched_init_bore(void) {
 	init_task.se.burst_time = 0;
